// src/types/application-event/INotificationMessage.d.ts
/*
 * This file is part of RootDB.
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as published
 * by the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU Affero General Public License for more details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 *
 * AUTHORS
 * PORQUET Sébastien <sebastien.porquet@ijaz.fr>
 * ROBIN Brice <brice@robri.net>
 */

import IApplicationMessage  from "./IApplicationMessage";
import { TSeverity }        from "../common/TSeverity";
import IReportDevBarMessage from "./IReportDevBarMessage";

interface INotificationMessage extends IApplicationMessage, IReportDevBarMessage {
    severity: TToastSeverity,
    toast: boolean,
    buttonLabel?: string,
    buttonSeverity?: TSeverity
    buttonURL?: string,
    buttonURLInNewTab?: boolean
    life?: number,
    new?: boolean,
}

export = INotificationMessage;